No need to split job type checking after all.
